chore: 타입에 맞게 코드 분리 및 파일명 변경
- CalulateItem을 채택한 Operator, Double을 CalculateItem 파일에 묶음
